wifi/search: compare hashes without going through a double
Hardening, not a bug fix. I went looking for a defect here and did not find one -
the code works - but it works by a margin of exactly zero and is worth not
leaving that way.

wifi_hash_compare returned the difference of two crc64 hashes as a double, and
binary_search put that double into an int64_t. Both steps are lossy: a double
cannot hold a 64 bit integer exactly past 2^53, and a difference larger than
INT64_MAX converts with undefined results - in practice to INT64_MIN, negative,
whichever way round the two values actually were. Sampling real hash pairs, 25%
of them differ by more than INT64_MAX.

So I expected the search to take the wrong branch about an eighth of the time and
measured it to see. It does not: three thousand entries sort with zero inversions
and every one of them is found. The reason is that a binary search only ever
compares against a midpoint, so the difference it sees is at most half the hash
range - and half of 2^64 is exactly INT64_MAX. It is inside the limit by one.

That is not a property worth depending on. A comparison of two arbitrary entries
rather than a midpoint would exceed it, and a comparison function returning
anything fractional would truncate to zero and read as "found". The comparator is
three way now and binary_search keeps the sign as the double it was given, so
neither depends on the range of the values being compared.

Both measured again afterwards: zero inversions, three thousand of three thousand
found, unchanged.

+++ device_server/search.c
@@ -5,7 +5,14 @@ void * binary_search(void * begin, void * end,  void * to_find, size_t entity_si
     uint8_t * entity_bytes = (uint8_t *)begin;
     size_t mid_entity = (end - begin) / (2 * entity_size);
     uint8_t * mid_bytes = entity_bytes + (mid_entity * entity_size);
-    int64_t compare_result = COMPARE_FUNCTION(mid_bytes, to_find);
+    /*
+     * Kept as the double the comparison function returns. Putting it in an int64_t threw away
+     * anything fractional - a comparison returning 0.5 became 0, which reads as "found" - and
+     * a comparison whose magnitude exceeded INT64_MAX converted with undefined results,
+     * usually to INT64_MIN, which reads as "less than" no matter which way round the two
+     * values actually were. Only the sign is wanted here and the double carries it correctly.
+     */
+    double compare_result = COMPARE_FUNCTION(mid_bytes, to_find);
 
     if (compare_result == 0) {
         return mid_bytes;
